\encoding{UTF-8}
\name{occupy-package}
\Rdversion{1.1}
\alias{occupy-package}
\alias{occupy}
\docType{package}
\title{
occupy: Analyzing Single Visit Site Occupancy Data with Detection Error
}
\description{
The package implements zero inflated binomial models
  for single visit site occupancy data when the probability of detection
  is less than one. Provides hands-on formula interface and model selection
  tools based on quasi-Newton method and data cloning.
  Development was funded by the Alberta Biodiversity Monitoring Institute
  \url{http://www.abmi.ca}.
}
\details{
\code{\link{singleocc}}: easy model fitting.
}
\author{
\enc{P\'eter S\'olymos}{Peter Solymos} and Monica Moreno

Maintainer: \enc{P\'eter S\'olymos}{Peter Solymos}, \email{solymos@ualberta.ca}
}
\references{
REFS HERE
}
\keyword{ package }
